Make download asynchronous
There is no method to know when a download is complete but there is
a way to make the API to do so async. This makes the opportunity to
handle cleanup code even though the actual act of downloading happens
outside the runtime event loop.

This change adds a small setTimeout delay to simulate an asynchronous
resolution. It isn't perfect but it is enough to allow cleanup code.

The specific use case that drove this change was the idea that an iframe
could load an ember-quine app and safely remove that iframe after the
download "completed". In this case it simulates a resolution which is
enough for the browser to open the save file dialog while the iframe is
removed under the hood. Once the save file dialog is shown the data it
saves is immutable and won't disappear till after the dialog closes even
if the source (the iframe) who opened the dialog via the click() is no
longer there.