Insert document store area to top of DOM
Prior to this change, the document store area was being placed at the
bottom of the body. Unfortunately I discovered that Ember boots
immediately and does not wait for DOMContentLoaded which means the
document store area was unavailable to the app till after it completed
booting.

This change inserts the document store area to the top of the body tag
so that it is available as the Ember app boots.