Use better filtering for cloneBody
Prior to this change, the cloneBody was taking all DOM and filtering out
children with a class of ember-view. Unfortunately Glimmer components do
not offer the same class and so this is a poor filter to use.

This change captures only DOM children that are expected to be cloned.
This moves from an opt-out (in the case of ember-view) to an opt-in
mechanism.

By default the DOM nodes that ember-quine produces (document store area
and the style/scripts that get embedded (marked with data-asset-path
attribute) are included and also custom child nodes with the data-quine
attribute if desired.